refactor: move session-detail and timeline
Move components/session-detail and components/timeline to nerdlets/video-session/components
