+++ nerdlets/video-session/components/session-detail/styles.scss
