+++ nerdlets/video-session/components/timeline/EventStream.js
