+++ nerdlets/video-session/components/timeline/VideoGroup.js
