+++ nerdlets/video-session/components/timeline/gauge/Gauge.js
