+++ nerdlets/video-session/components/timeline/gauge/GaugeDataLegendItem.js
