+++ nerdlets/video-session/components/timeline/gauge/display-helpers.js
