+++ nerdlets/video-session/components/timeline/gauge/styles.scss
