+++ nerdlets/video-session/components/timeline/styles.scss
