+++ nerdlets/video-session/styles.scss
@@ -1,5 +1,5 @@
-@import '../../components/session-detail/styles';
-@import '../../components/timeline/styles';
+@import './components/session-detail/styles';
+@import './components/timeline/styles';
 
 .column {
   background-color: white;
